Lazy load all API definition widgets
